annotations

+++ reportly-common-domain-model/src/main/java/fr/isika/cda/entities/users/Employee.java
@@ -16,7 +16,7 @@ public class Employee extends UserAccount implements Serializable {
 	 */
 	private static final long serialVersionUID = 6240533294536565573L;
 
-	private int tjm;
+	private double tjm;
 	private boolean busy;
 	private String skills;
